Base: Rufs update
* ColorPicker, MessageInf, PrepareSentence, Session, Rondutils

+++ kriolos-opos-base/src/main/java/com/openbravo/pos/util/RoundUtils.java
@@ -56,6 +56,6 @@ public class RoundUtils {
      * @return
      */
     public static double getValue(Double value) {
-        return value == null ? 0.0 : value.doubleValue();
+        return value == null ? 0.0 : value;
     }
 }
